requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow>=1.3.0,<2.0.0
keras==2.2.5
opencv-python
h5py
imgaug
IPython[all]